拆分utils
